do not traverse standard library dependencies
The set of packages dependended on by the standard
library changes from release to release. In general
we want a go.dep file to be good to work over several
releases at the same time, so we stop traversing
dependencies when we get to the standard library.

This should work OK as a heuristic because the standard library
already controls dependencies well.