remove stdtest concept
We don't really want to use any package specifiers that
the go command does not support, so remove the `stdtest` concept.
Instead, we treat test packages within the standard library as special - they
are not matched by `std` but instead must be mentioned specifically.

This makes it possible to allow all standard library dependencies with
`std` with the exception of the `testing` package.

In the future, this might be replaced by some more general
exclusion mechanism, but this seems like a reasonable compromise for now.